Refactored Classes and project Structure
