--- src/main/java/com/example/capstone/model/Person.java
@@ -1,28 +1,0 @@
-package com.example.capstone.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import javax.validation.constraints.NotBlank;
-import java.util.UUID;
-
-
-public class Person {
-
-    private final UUID id;
-    @NotBlank
-    private final String name;
-
-    public Person(@JsonProperty("id") UUID id,
-                  @JsonProperty("name") String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
