--- src/main/java/com/example/capstone/service/PersonService.java
@@ -1,42 +1,0 @@
-package com.example.capstone.service;
-
-import com.example.capstone.dao.PersonDao;
-import com.example.capstone.model.Person;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@Service
-public class PersonService {
-
-    private final PersonDao personDao;
-
-    @Autowired
-    public PersonService(@Qualifier("mysql1") PersonDao personDao) {
-        this.personDao = personDao;
-    }
-
-    public int addPerson(Person person) {
-        return personDao.insertPerson(person);
-    }
-
-    public List<Person> getAllPeople() {
-        return personDao.selectAllPeople();
-    }
-
-    public Optional<Person> getPersonById(UUID id) {
-        return personDao.selectPersonById(id);
-    }
-
-    public int deletePerson(UUID id) {
-        return personDao.deletePersonById(id);
-    }
-
-    public int updatePerson(UUID id, Person newPerson) {
-        return personDao.updatePersonById(id, newPerson);
-    }
-}
