+++ src/main/java/com/project/capstone/CapstoneApplication.java
@@ -1,0 +1,16 @@
+package com.project.capstone;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CapstoneApplication {
+
+
+	public static void main(String[] args) {
+		SpringApplication.run(CapstoneApplication.class, args);
+	}
+
+}
+
+
