+++ src/main/java/com/project/capstone/personexample/Person.java
@@ -1,0 +1,28 @@
+package com.project.capstone.personexample;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import javax.validation.constraints.NotBlank;
+import java.util.UUID;
+
+
+public class Person {
+
+    private final UUID id;
+    @NotBlank
+    private final String name;
+
+    public Person(@JsonProperty("id") UUID id,
+                  @JsonProperty("name") String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
