+++ src/main/java/com/project/capstone/personexample/PersonController.java
@@ -1,0 +1,47 @@
+package com.project.capstone.personexample;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.NonNull;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("api/v1/person1")
+@RestController
+public class PersonController {
+
+    private final PersonService personService;
+
+
+    @Autowired
+    public PersonController(PersonService personService) {
+        this.personService = personService;
+    }
+
+    @PostMapping
+    public void addPerson(@Valid @NonNull @RequestBody Person person) {
+        personService.addPerson(person);
+    }
+
+    @GetMapping
+    public List<Person> getAllPeople() {
+        return personService.getAllPeople();
+    }
+
+    @GetMapping(path = "{id}")
+    public Person getPersonById(@PathVariable("id") UUID id) {
+        return personService.getPersonById(id).
+                orElse(null);
+    }
+    @DeleteMapping(path = "{id}")
+    public void deletePersonById(@PathVariable("id") UUID id) {
+        personService.deletePerson(id);
+    }
+
+    @PutMapping(path = "{id}")
+    public void updatePerson(@PathVariable("id") UUID id, @Valid @NonNull @RequestBody Person personToUpdate) {
+        personService.updatePerson(id, personToUpdate);
+    }
+}
