+++ src/main/java/com/project/capstone/personexample/PersonDao.java
@@ -1,0 +1,24 @@
+package com.project.capstone.personexample;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface PersonDao {
+    int insertPerson(UUID id, Person person);
+
+    default int insertPerson(Person person) {
+        UUID id = UUID.randomUUID();
+        return insertPerson(id, person);
+    }
+
+    List<Person> selectAllPeople();
+
+    Optional<Person> selectPersonById(UUID id);
+
+    int deletePersonById(UUID id);
+
+    int updatePersonById(UUID id, Person person);
+
+
+}
