+++ src/main/java/com/project/capstone/user/UserDao.java
@@ -1,0 +1,19 @@
+package com.project.capstone.user;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserDao {
+
+    int insertUser(User user);
+
+
+    List<User> selectAllUsers();
+
+    Optional<User> selectUserById(int id);
+
+    int deleteUserById(int id);
+
+    int updateUserById(int id, User user);
+
+}
