+++ src/test/java/com/project/capstone/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.capstone;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
